+++ changelog.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+#### KMD Nova
+
+- Department and Caseworker classes added.
+- Ability to set security unit, department and caseworker on cases.
+- Better tests for cases.
+
+### Fixed
+
+- Security unit not set properly on Nova cases.
+
 ## [1.3.1] - 2024-02-21
 
 ### Fixed